refactor: put formatDataForAPI in its own file + refactor types
