+++ src/globals.ts
@@ -35,20 +35,20 @@ export type AppState = {
   indicateurDeuxTrois: {
     formValidated: FormState
     presenceAugmentationPromotion: boolean
-    nombreAugmentationPromotionFemmes: number | undefined
-    nombreAugmentationPromotionHommes: number | undefined
+    nombreAugmentationPromotionFemmes?: number | undefined
+    nombreAugmentationPromotionHommes?: number | undefined
     periodeDeclaration: PeriodeDeclaration
   } & Partial<DeclarationIndicateurDeuxTroisData>
   indicateurQuatre: {
     formValidated: FormState
     presenceCongeMat: boolean
-    nombreSalarieesPeriodeAugmentation: number | undefined
-    nombreSalarieesAugmentees: number | undefined
+    nombreSalarieesPeriodeAugmentation?: number | undefined
+    nombreSalarieesAugmentees?: number | undefined
   } & Partial<DeclarationIndicateurQuatreData>
   indicateurCinq: {
     formValidated: FormState
-    nombreSalariesHommes: number | undefined
-    nombreSalariesFemmes: number | undefined
+    nombreSalariesHommes?: number | undefined
+    nombreSalariesFemmes?: number | undefined
   } & Partial<DeclarationIndicateurCinqData>
   informationsEntreprise: {
     formValidated: FormState
@@ -63,7 +63,7 @@ export type AppState = {
     commune: string
     structure: Structure
     nomUES: string
-    nombreEntreprises: number | undefined
+    nombreEntreprises?: number | undefined
     entreprisesUES: Array<EntrepriseUES>
   }
   informationsDeclarant: {
@@ -77,15 +77,15 @@ export type AppState = {
   declaration: {
     formValidated: FormState
     mesuresCorrection: string
-    cseMisEnPlace: boolean | undefined
+    cseMisEnPlace?: boolean | undefined
     dateConsultationCSE: string
     datePublication: string
-    publicationSurSiteInternet: boolean | undefined // uniquement un boolean quand une déclaration est valide
+    publicationSurSiteInternet?: boolean | undefined // uniquement un boolean quand une déclaration est valide
     lienPublication: string
     planRelance: boolean | undefined // uniquement un boolean quand une déclaration est valide
     modalitesPublication: string
     dateDeclaration: string
-    noteIndex: number | undefined
+    noteIndex?: number | undefined
     totalPoint: number
     totalPointCalculable: number
   }
@@ -428,8 +428,8 @@ export enum CategorieSocioPro {
 
 export interface GroupTranchesAgesEffectif {
   trancheAge: TranchesAges
-  nombreSalariesFemmes: number | undefined
-  nombreSalariesHommes: number | undefined
+  nombreSalariesFemmes?: number | undefined
+  nombreSalariesHommes?: number | undefined
 }
 
 export interface GroupeEffectif {
@@ -439,8 +439,8 @@ export interface GroupeEffectif {
 
 export interface GroupTranchesAgesIndicateurUn {
   trancheAge: TranchesAges
-  remunerationAnnuelleBrutFemmes: number | undefined
-  remunerationAnnuelleBrutHommes: number | undefined
+  remunerationAnnuelleBrutFemmes?: number | undefined
+  remunerationAnnuelleBrutHommes?: number | undefined
   ecartTauxRemuneration?: number | undefined
 }
 
@@ -465,15 +465,15 @@ export interface GroupeCoefficient {
 
 export interface GroupeIndicateurDeux {
   categorieSocioPro: CategorieSocioPro
-  tauxAugmentationFemmes: number | undefined
-  tauxAugmentationHommes: number | undefined
+  tauxAugmentationFemmes?: number | undefined
+  tauxAugmentationHommes?: number | undefined
   ecartTauxAugmentation?: number | undefined
 }
 
 export interface GroupeIndicateurTrois {
   categorieSocioPro: CategorieSocioPro
-  tauxPromotionFemmes: number | undefined
-  tauxPromotionHommes: number | undefined
+  tauxPromotionFemmes?: number | undefined
+  tauxPromotionHommes?: number | undefined
   ecartTauxPromotion?: number | undefined
 }
